# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Scorekeeper API
  description: |
    API for the Scorekeeper word puzzle game.

    See also: [Planned Features](https://hannahscovill.github.io/scorekeeper/planned)
  version: 0.1.0
servers:
  - url: https://api.wordles.dev
    description: Remote
  - url: http://localhost:8080
    description: Local
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
...
  /profile:
    get:
      summary: Get user profile
      description: |
        Retrieve the authenticated user's profile from Auth0 user_metadata.

        Requires Auth0 M2M credentials to be configured on the server.
      operationId: getProfile
      tags:
        - Users
      responses:
        "200":
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: "schemas.yaml#/ProfileResponse"
              examples:
                default:
                  $ref: "data-examples/examples.yaml#/ProfileResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      summary: Update user profile
      description: |
        Update the authenticated user's profile in Auth0 user_metadata.

        Changes will be reflected in new ID tokens issued after the update.
        Requires Auth0 M2M credentials to be configured on the server.
      operationId: updateProfile
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "schemas.yaml#/UpdateProfileRequest"
            examples:
              default:
                $ref: "data-examples/examples.yaml#/UpdateProfileRequest"
      responses:
        "200":
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: "schemas.yaml#/ProfileResponse"
              examples:
                default:
                  $ref: "data-examples/examples.yaml#/ProfileResponseUpdated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /profile/avatar:
    post:
      summary: Upload user avatar
      description: |
        Upload an avatar image for the authenticated user.

        The image is uploaded to S3 and the URL is stored in Auth0 user_metadata.
        Requires both Auth0 M2M credentials and S3_AVATAR_BUCKET to be configured.
      operationId: uploadAvatar
      tags:
        - Users
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - avatar
              properties:
                avatar:
                  type: string
                  format: binary
                  description: Image file to upload (JPEG or PNG, max 2MB)
      responses:
        "200":
          description: Avatar uploaded successfully
          content:
            application/json:
              schema:
                $ref: "schemas.yaml#/AvatarUploadResponse"
              examples:
                default:
                  $ref: "data-examples/examples.yaml#/AvatarUploadResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          description: Internal Server Error (S3 upload failure)
          content:
            application/json:
              schema:
                $ref: "schemas.yaml#/ErrorResponse"
              example:
                message: "Failed to upload avatar"
  /profile/mobile-test-track-signup:
    post:
      summary: Opt into the internal mobile test track (authenticated)
      description: |
        Opts the authenticated user into the iOS and/or Android internal test
        track, writing directly to their Auth0 user_metadata.

        Idempotent per platform — resubmitting a platform that's already
        opted in is a no-op (no Auth0 write, no GitHub notification).

        Notifies via a GitHub issue (label `mobile-test-user-opt-in`)
        containing the user's display name and Auth0 user ID only — the
        email is never included, even though it's a field on the account.
        GitHub notification is best-effort: if GitHub credentials aren't
        configured, the opt-in still succeeds and the notification is
        silently skipped.

        **No Turnstile token required** — a valid JWT already proves the
        request is from a real logged-in user, so CAPTCHA verification
        (required on the anonymous `POST /mobile-test-track-signup` below) would be
        redundant here.
      operationId: mobileTestTrackSignup
      tags:
        - Mobile Test Track Signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "schemas.yaml#/MobileTestTrackSignupRequest"
            examples:
              default:
                $ref: "data-examples/examples.yaml#/MobileTestTrackSignupRequest"
      responses:
        "200":
          description: Current opt-in state after the call
          content:
            application/json:
              schema:
                $ref: "schemas.yaml#/MobileTestTrackSignupResponse"
              examples:
                default:
                  $ref: "data-examples/examples.yaml#/MobileTestTrackSignupResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token in Authorization header
    cookieAuth:
      type: apiKey
      in: cookie
      name: wordle_session
      description: JWT token in wordle_session cookie
  responses:
...
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "schemas.yaml#/ErrorResponse"
          example:
            message: "bad request"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "schemas.yaml#/ErrorResponse"
          example:
            message: "unauthorized"
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "schemas.yaml#/ErrorResponse"
          example:
            message: "not found"
...